separated the two options into separated file + fix a bug with settings being below study modal
