+++ src/components/Verse/OverflowVerseActionsMenuBody/index.tsx
@@ -1,23 +1,19 @@
 import React, { useState } from 'react';
 
-import useTranslation from 'next-translate/useTranslation';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import PinVerseAction from '../PinVerseAction';
 import TranslationFeedbackAction from '../TranslationFeedback/TranslationFeedbackAction';
 import VerseActionAdvancedCopy from '../VerseActionAdvancedCopy';
 import VerseActionEmbedWidget from '../VerseActionEmbedWidget';
 import VerseActionRepeatAudio from '../VerseActionRepeatAudio';
+import VerseActionSettings from '../VerseActionSettings';
+import VerseActionTranslationsSettings from '../VerseActionTranslationsSettings';
 
 import ShareVerseActionsMenu from './ShareVerseActionsMenu';
 
 import VerseActionsMenuType from '@/components/QuranReader/ReadingView/WordActionsMenu/types';
 import WordByWordVerseAction from '@/components/QuranReader/ReadingView/WordByWordVerseAction';
-import IconContainer, { IconColor, IconSize } from '@/dls/IconContainer/IconContainer';
-import PopoverMenu from '@/dls/PopoverMenu/PopoverMenu';
-import SettingsIcon from '@/icons/settings.svg';
-import TranslationsIcon from '@/icons/translation.svg';
-import { setIsSettingsDrawerOpen, setSettingsView, SettingsView } from '@/redux/slices/navbar';
 import { selectStudyModeIsOpen } from '@/redux/slices/QuranReader/studyMode';
 import Verse from 'types/Verse';
 
@@ -34,25 +30,8 @@ const OverflowVerseActionsMenuBody: React.FC<Props> = ({
   onActionTriggered,
   isInsideStudyMode = false,
 }) => {
-  const { t } = useTranslation('common');
-  const dispatch = useDispatch();
-
   const [selectedMenu, setSelectedMenu] = useState<VerseActionsMenuType>(VerseActionsMenuType.Main);
   const isStudyModeOpen = useSelector(selectStudyModeIsOpen);
-
-  // Open the translations settings
-  const onTranslationsSettingsClick = () => {
-    dispatch(setSettingsView(SettingsView.Translation));
-    dispatch(setIsSettingsDrawerOpen(true));
-    onActionTriggered?.();
-  };
-
-  // Open the general settings
-  const onSettingsClick = () => {
-    dispatch(setSettingsView(SettingsView.Body));
-    dispatch(setIsSettingsDrawerOpen(true));
-    onActionTriggered?.();
-  };
 
   return selectedMenu === VerseActionsMenuType.Main ? (
     <div>
@@ -73,20 +52,7 @@ const OverflowVerseActionsMenuBody: React.FC<Props> = ({
       {!isStudyModeOpen && (
         <VerseActionRepeatAudio isTranslationView={isTranslationView} verseKey={verse.verseKey} />
       )}
-      <PopoverMenu.Item
-        icon={
-          <IconContainer
-            icon={<TranslationsIcon />}
-            color={IconColor.tertiary}
-            size={IconSize.Custom}
-            shouldFlipOnRTL={false}
-          />
-        }
-        onClick={onTranslationsSettingsClick}
-        shouldCloseMenuAfterClick
-      >
-        {t('translations')}
-      </PopoverMenu.Item>
+      <VerseActionTranslationsSettings onActionTriggered={onActionTriggered} />
       <TranslationFeedbackAction
         verse={verse}
         isTranslationView={isTranslationView}
@@ -98,20 +64,7 @@ const OverflowVerseActionsMenuBody: React.FC<Props> = ({
         isTranslationView={isTranslationView}
         onActionTriggered={onActionTriggered}
       />
-      <PopoverMenu.Item
-        icon={
-          <IconContainer
-            icon={<SettingsIcon />}
-            color={IconColor.tertiary}
-            size={IconSize.Custom}
-            shouldFlipOnRTL={false}
-          />
-        }
-        onClick={onSettingsClick}
-        shouldCloseMenuAfterClick
-      >
-        {t('settings.title')}
-      </PopoverMenu.Item>
+      <VerseActionSettings onActionTriggered={onActionTriggered} />
     </div>
   ) : (
     <ShareVerseActionsMenu
